requirements.txt
fastapi==0.85.0
uvicorn==0.13.4
fastapi-sqlalchemy==0.2.1
python-dotenv==0.15.0
streamlit==1.36.0
underthesea==6.8.4
transformers
tokenizers
torch
